requirements.txt
django
django-cors-headers
django-filter
django-mathfilters
django-model-utils
djangorestframework
python-decouple
psycopg2
markdown
